pyproject.toml
[project]
name = "gpsimagestomap"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.3",
    "gpxpy>=1.6.2",
    "piexif>=1.1.3",
    "pillow>=12.2.0",
    "pillow-heif>=1.3.0",
]

[project.scripts]
gpsimagestomap = "gpsimagestomap.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gpsimagestomap"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
